Rename main class and package
