+++ src/test/java/seedu/contech/ConTechTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.contech;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
